update: move chat settings and perms to chat interface
